PauseTiming and ResumeTiming are now private.
Use fixtures if you need to do significant work to prepare for the
benchmark to run.

Fixes #179.
